p9 scom refactor to support multiple platforms at runtime
